refact: move views_workflow into chord.workflows submodule
